Adiciona pasta no padrao da arquitetura e move arquivos de comando
