+++ template/backend/src/Ambev.DeveloperEvaluation.Application/Products/CreateProduct/CreateProductCommandValidator.cs
@@ -1,6 +1,6 @@
 ﻿using FluentValidation;
 
-namespace Ambev.DeveloperEvaluation.Application.Products
+namespace Ambev.DeveloperEvaluation.Application.Products.CreateProduct
 {
     /// <summary>
     /// Validator for CreateProductCommand that defines validation rules for product creation command.
