+++ template/backend/src/Ambev.DeveloperEvaluation.Application/Products/CreateProduct/CreateProductProfile.cs
@@ -1,0 +1,20 @@
+﻿using Ambev.DeveloperEvaluation.Domain.Entities;
+using AutoMapper;
+
+namespace Ambev.DeveloperEvaluation.Application.Products.CreateProduct
+{
+    /// <summary>
+    /// Profile for mapping between Product entity and CreateProductResponse
+    /// </summary>
+    public class CreateProductProfile : Profile
+    {
+        /// <summary>
+        /// Initializes the mappings for CreateProduct feature
+        /// </summary>
+        public CreateProductProfile()
+        {
+            CreateMap<CreateProductCommand, Product>();
+            CreateMap<Product, CreateProductResult>();
+        }
+    }
+}
