+++ template/backend/src/Ambev.DeveloperEvaluation.Application/Products/CreateProduct/CreateProductResult.cs
@@ -1,4 +1,4 @@
-﻿namespace Ambev.DeveloperEvaluation.Application.Products
+﻿namespace Ambev.DeveloperEvaluation.Application.Products.CreateProduct
 {
     /// <summary>
     /// Represents the response returned after successfully creating a new product.
@@ -14,5 +14,19 @@
         /// </summary>
         /// <value>A GUID that uniquely identifies the created product in the system.</value>
         public Guid Id { get; set; }
+
+        /// <summary>
+        /// Gets or sets the products's full name.
+        /// Must not be null or empty.
+        /// </summary>
+        public string Name { get; set; } = string.Empty;
+        /// <summary>
+        /// Gets or sets the description of product.
+        /// </summary>
+        public string Description { get; set; } = string.Empty;
+        /// <summary>
+        /// Gets or sets the price of product.
+        /// </summary>
+        public decimal Price { get; set; }
     }
 }
